@@ -11,7 +11,7 @@
   ],
   "require": {
     "guzzlehttp/guzzle": "^6.3",
-    "php":"^7.1",
+    "php":"^7.4",
     "ext-mbstring": "*"
   },
   "require-dev": {
