@@ -3,26 +3,15 @@ declare(strict_types=1);
 
 namespace Comgate\Request;
 
+use Comgate\Response\BaseResponse;
+
 interface RequestInterface
 {
-
-    /**
-     * @return array
-     */
     public function getData(): array;
 
-
-    /**
-     * @return bool
-     */
     public function isPost(): bool;
 
-
-    /**
-     * @return string
-     */
     public function getEndPoint(): string;
 
-
-    public function getResponseObject(array $data);
+    public function getResponseObject(string $data): BaseResponse;
 }
