@@ -3,6 +3,6 @@ declare(strict_types=1);
 
 namespace Comgate\Response;
 
-class RefundPaymentResponse extends BaseResponse
+class RefundPayment extends BaseResponse
 {
 }
